move topology/allocate_volume.go to operation/allocate_volume.go
